requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
openai-whisper==20250625
PySide6==6.10.1
transformers==4.57.3
sentencepiece==0.2.1
srt==3.5.3
rich==14.2.0
numba==0.63.0
llvmlite==0.46.0
tiktoken==0.12.0
requests==2.32.5
